28082019 b

+++ app/src/main/java/com/empoderar/picar/model/persistent/database/data/TypeFormData.kt
@@ -8,15 +8,15 @@ import android.arch.persistence.room.PrimaryKey
 data class TypeFormData(@PrimaryKey var frmID: String,
                         @ColumnInfo(name = "idStep") var idStep: String,
                         @ColumnInfo(name = "title") var title: String,
-                        @ColumnInfo(name = "max") var max: Float,
-                        @ColumnInfo(name = "weigh") var weigh: Float,
+                        @ColumnInfo(name = "max") var max: Double,
+                        @ColumnInfo(name = "weigh") var weigh: Double,
                         @ColumnInfo(name = "observation") var observation: String,
                         @ColumnInfo(name = "maxBit") var maxBit: Boolean,
                         @ColumnInfo(name = "valueBit") var valueBit: Boolean,
                         @ColumnInfo(name = "logicBit") var logicBit: Boolean,
                         @ColumnInfo(name = "dateBit") var dateBit: Boolean,
                         @ColumnInfo(name = "textBit") var textBit: Boolean) {
-    constructor(): this("", "", "", 0.0f, 0.0f,
+    constructor(): this("", "", "", 0.0, 0.0,
             "", false, false, false,
             false, false)
 }
